+++ openghg/analyse/_scenario.py
@@ -180,8 +180,12 @@ class ModelScenario:
         TODO: For obs, footprint, flux should we also allow Dataset input and turn
         these into the appropriate class?
         """
+        self.platform: str | None = platform
         self.obs: ObsData | ObsColumnData | None = None
         self.obs_column: ObsColumnData | None = None
+        if self.obs_column is not None:
+            if platform is not None:
+                self.platform = platform
         self.footprint: FootprintData | None = None
         self.fluxes: dict[str, FluxData] | None = None
         self.bc: BoundaryConditionsData | None = None
@@ -193,7 +197,7 @@ class ModelScenario:
 
         if satellite is not None and platform is None:
             logger.info("You passed a satellite but no platform. Updating the platform to 'satellite'")
-            platform = "satellite"
+            self.platform = "satellite"
 
         # For ObsColumn data processing
         if platform in accepted_column_data_types:
@@ -242,11 +246,11 @@ class ModelScenario:
         self.add_footprint(
             site=site,
             inlet=inlet,
+            satellite=satellite,
             height=height,
             domain=domain,
             model=model,
             met_model=met_model,
-            satellite=satellite,
             fp_inlet=fp_inlet,
             start_date=start_date,
             end_date=end_date,
@@ -470,7 +474,8 @@ class ModelScenario:
                 if self.obs is not None:
                     network = network or self.obs.metadata.get("network")
 
-                height_name = extract_height_name(site, network, inlet)
+                if site is not None:
+                    height_name = extract_height_name(site, network, inlet)
                 if height_name is not None:
                     fp_inlet = height_name
                     logger.info(
@@ -753,7 +758,7 @@ class ModelScenario:
             # Do not apply resampling for "satellite" or "flask"
             if platform == "satellite":
                 resample_to = None
-                align_to_obs = False
+                align_to_obs = True
             elif "flask" in platform:
                 resample_to = None
                 align_to_obs = True
@@ -762,7 +767,13 @@ class ModelScenario:
 
         if resample_to is None:
             if align_to_obs:
-                footprint_data = footprint_data.reindex_like(obs_data, method="ffill")
+                if platform == "satellite":
+                    footprint_data = footprint_data.reindex_like(
+                        obs_data, method="nearest", tolerance=pd.Timedelta("1ms")
+                    )
+                    logger.info("Reindexing footprint data to satellite observation data.")
+                else:
+                    footprint_data = footprint_data.reindex_like(obs_data, method="ffill")
             return obs_data, footprint_data
 
         if resample_to == "footprint":
@@ -978,7 +989,7 @@ class ModelScenario:
         # if self.modelled_obs is None or recalculate:
         if parameter is None or recalculate:
             # Check if observations are present and use these for resampling
-            if self.obs is not None:
+            if self.obs is not None or self.obs_column is not None:
                 self.combine_obs_footprint(
                     resample_to, platform=platform, recalculate=recalculate, cache=True
                 )